requirements.txt
Flask
Flask_SQLAlchemy
PyYAML
Werkzeug
